[rename] runnables -> kernels
