+++ lambda/src/kernels/mod.rs
