+++ lambda/src/lib.rs
@@ -1,3 +1,3 @@
 pub mod components;
 pub mod core;
-pub mod runnables;
+pub mod kernels;
