+++ lambda/tests/runnables.rs
@@ -2,5 +2,5 @@ use lambda;
 
 #[test]
 fn lambda_runnable() {
-  use lambda::runnables::LambdaKernel;
+  use lambda::kernels::LambdaKernel;
 }
